Bug 1126607, part 1: Import the LDAP C-SDK into comm-central.
This code is generated by finding all .c, .h, and relevent Makefile.in files in
the relevant files of c-sdk/ldap from http://hg.mozilla.org/projects/ldap-sdks/,
as of LDAPCSDK_6_0_7G_RTM. The file directories are re-mapped in the import to
avoid needing to do extensive clobbering of the source and build directories.

No other changes are done in this changeset, and the imported code is not
integrated into the build system.