Bug 1295089 - Remove unneeded 'status-text' statusbarpanel; r=MakeMyDay, ui-r=Paenglab
A 'status-text' statusbarpanel was added to the calendar
statusbar in bug 1277972. (This was because there was a
similar statusbarpanel in the window dialog for editing
events/tasks.) But there is no need for it since the
statusbar already has a statusbarpanel for this. Removing
it fixes some spacing problems in the statusbar.