Bug 1456589 - Part 4 Port Bug 1434376 and Bug 1432425 to SeaMonkey places. r=IanN a=IanN
Port Bug 1434376 [Switch over all uses of BrowserUtils.promiseLayoutFlushed to window.promiseDocumentFlushed].
Comment only for suite.

window.promiseDocumentFlushed will call a callback as soon as a style or layout
flush is not required for the document (which might be immediately). This is a
new ChromeOnly API introduced in an earlier patch in this series.

Port Bug 1432425 [Remove synchronous Bookmarks::RemoveFolderChildren].