Bug 1547631 - Port bug 1523684: Set 32-bit build target on win32 l10n repacks. r=jorgk
L10n repacks happen on 64-bit hosts since bug 1522354. Therefore --target
needs to be explicitly set to i686-pc-mingw32 on l10n builds otherwise it
defaults to x86_64-pc-mingw32 which produces a 64-bit installer.

The contents of the installer are 32-bit as --target is set in
build/win32/mozconfig.vs2017.