Bug 1634209 - Build librnp on Linux statically linking to LLVM's libc++. r=kaie a=rjl
Circumvents the problems with libstdc++ symbols problem by replacing it with
LLVM's libc++ only when building on Taskcluster. libc++abi is a required
companion library.
Since libc++ is not as ubiquitous as libstdc++ on Linux, it is statically linked.
The libc++ license permits this quite explicitly.
When librnp (and more precisely Botan) is compiled and linked this way, there are
no references to libstdc++ in the resulting librnp.so file.

Depends on D80223

Differential Revision: https://phabricator.services.mozilla.com/D80224