Bug 1670832 - Upgrade browser_ext_menus.js some more. r=mkmelin
I'm improving this test to prove that the menu in the compose window does the same things as the main context menu, with respect to WebExtensions.
More big changes to the menus code are in the pipeline as well, so locking down the behaviour with a test is good.

Adds tests for the selection, image, and link contexts. Adds tests for other places the main context menu appears.

Differential Revision: https://phabricator.services.mozilla.com/D94213