Bug 1677735 - Catch uncaught but expected NS_ERROR_BLOCKED_BY_POLICY in a test. rs=bustage-fix
I'm not entirely sure why this behaves differently on Firefox but it does. It seems to me the exception should throw for them too.
I think it has something to do with the relationship between their browser elements and gBrowser. The same behaviour occurs whether E10s is enabled or not.