Bug 1694631 - Stop using Lookup in CalendarUtils. r=darktrojan
The lookup and lookupEventBox methods require full tree paths to find elements, which makes them hard to maintain when restructuring.

These were dropped from CalendarUtils, along with the path related methods and constants.

The standard querySelector and getElementById methods were used instead.

In particular, the CalendarTestUtils module was expanded with the dayView, weekView and multiweekView submodules to replace the lookupEventBox usage.

Differential Revision: https://phabricator.services.mozilla.com/D107005