Bug 1694502 - Don't autocomplete and itemize recipients on blur if the window lost focus while editing them. r=aleca
When a blur event occurs on a row or pill input, until now we've always forced
autocompletion to the best match via input.forceComplete. However, if the user
switches focus to another window, we do *not* want force autocompletion so as to
allow further address editing, like pasting in a display name etc. So we're now
using input._dontBlur, which effectively eliminates toolkit's blur handling and
allows us to do our own blur handling, where we autcomplete on blur only when
the focus remains within the composition window.