Bug 1516423 - Have WinUnreadBadge own TaskbarWindowPreview.  r=mkmelin
In Thunderbird, an instance of `TaskbarWindowPreview` is created when
`WinUnreadBadge.updateUnreadCount` calls `WinTaskbar::GetOverlayIconController`.
However, that instance is freed immdiately after `updateUnreadCount` exits
because nobody owns it.  As a result, the taskbar icon disappears even when
`mail.minimizeToTray = false`.

This patch makes `WinUnreadBadge` own a reference to `TaskbarWindowPreview`
so that its instance is kept alive after the method exits.
This approach is the same as Firefox stores `TaskbarWindowPreview` as
`DownloadsTaskbar._taskbarProgress`.

Differential Revision: https://phabricator.services.mozilla.com/D110191