Bug 1668926 - fix missing member variable initialization for nsMsgMailboxParser::nsMsgMailboxParser. r=benc
uninit_member: Non-static class member m_startTime is not initialized in this constructor nor in any functions that it calls.
uninit_member: Non-static class member m_parsingDone is not initialized in this constructor nor in any functions that it calls.

CID 1260151 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)6. uninit_member: Non-static class member m_urlInProgress is not initialized in this constructor nor in any functions that it calls.

Differential Revision: https://phabricator.services.mozilla.com/D111621