Bug 1529929 - Allow filters to run even if filter logging fails. r=mkmelin
This makes the filter run a little more fault-tolerant. Previously, if
filterlog.html couldn't be written, the whole filter run would fail.
This changeset also takes the opportunity to tidy up a few of the rough
edges around the filter logging (eg closing the log file when the logging
is disabled).

Differential Revision: https://phabricator.services.mozilla.com/D113615