Bug 1683865 - Remove alt text that duplicates the title. r=mkmelin
Current specifications (see "Accessible Name and Description Computation" in HTML Accessibility API Mappings 1.0) state that if there is an alt attribute for an img it is used as the Accessible Name, otherwise the title attribute is used as the Accessible Name. See also Bug 1700174.

For existing situations where the title describes an icon for visual pointer users, and we also want this to be the Accessible Name, to avoid duplicating information, we should simply not include the alt attribute.

Note that these situations should still be avoided because a visual keyboard user would not have access to the title text.

Differential Revision: https://phabricator.services.mozilla.com/D116257