Bug 1721418 - [comm-beta] Run non-webrender test variant on Windows. r=justdave a=rjl
Intended to land directly to comm-beta, then ride the train to comm-esr91.

Creates a "noqr" (non-webrender) variant of the Windows webrender tests. This
is necessary because there is no "non-webrender" Windows test platform for us
to work with.
run-on-projects is adjusted so that the default (webrender) tests will not
run, but the noqa variant will.

Differential Revision: https://phabricator.services.mozilla.com/D120366