Bug 1679299 - Inline invitation-template.xhtml to prevent C++ re-entrancy in CalMimeConverter. r=darktrojan
This removes the use of a blocking (synchronous) XMLHttpRequest, which was
causing recursive calls to nsStreamConverter::OnDataAvailable() and general
mayhem. The blocking request was spinning the threads event queue, and so
the stream feeding the mime parsing sends us more data even though we're
still in the middle of processing the previous chunk!

Differential Revision: https://phabricator.services.mozilla.com/D120414