Bug 1737569 - Move calendar-event-column event metadata into eventDataMap. r=darktrojan a=wsmwk
We keep all the event metadata in one object, which is accessed through the event's hashId. This ensures each event has one set of data, which is easier to track.

Also, stopped using properties for some variables that are only used in one method.

Also, drop unused startMinute and endMinute properties.