Bug 1731729 - for manual account setup, when switching out TLS, use STARTTLS as default (not autodetect). r=aleca
When adjusting port and SSL was selected already, switch to StartTLS if the port is know to potentially use that.
After this, the user can still switch from StartTLS to None, but the initial selection was TLS, default to a something secure.

Fixes "typo" for pop3 preventing adjustments there to work correctly.
Use onchange instead of onblur for the change handling: when working with the spin buttons to change ports this is needed.

Differential Revision: https://phabricator.services.mozilla.com/D132353