Bug 1758399 - Strip symbols from libotr.dll when linking. r=#thunderbird-reviewers,aleca a=wsmwk
libotr.dll did not have debug symbols stripped prior to packaging due to not
really being integrated into the build. This resulted in Thunderbird shipping
an 11MB libotr.dll file.
Adding the "-s" argument to the link command brings the DLL size down to ~1.5MB
which is in-line with other platforms.
I verified that Linux and macOS builds of libotr do not have debug symbols in
the shipped builds.

Differential Revision: https://phabricator.services.mozilla.com/D140505