Bug 1755036 - Use a CSS class to indicate an element collapsed by a splitter. r=aleca
Instead of adjusting the styling of the collapsed element directly, we leave this decision to the stylesheet. This gives more flexibility.

In particular, it allows us to set a min-height and min-width on the 3pane elements that may collapse. Whilst this would normally be redundant, if the window is resized, this ensures that the elements will not shrink below their collapse sizes.

Differential Revision: https://phabricator.services.mozilla.com/D140499

Depends on D140498