Bug 1748449 - Introduce new widget for displaying invitations. r=aleca
This is the first step in re-designing the invitation display meant only to
lay down the basic structure and give a taste of the new widget. There is
still much work to do but doing it in one go is counter-productive.

Some tasks outstanding include:

1. Detecting the stage of the invitation and showing the appropriate UI.
2. Support for recurring events.
3. Toggle the timezone displayed.
4. Hiding the fields we do not have values for.
5. Improve the formatting of the attendees list.
6. Allow the user to hide long attendee lists.
7. Hooking up the action buttons.

Known issues (not exhaustive):
1. When there is a lot of information, the message pane gets pushed out of view.
2. When there is a lot of information, the widget is not as neat.
3. There are two scrollbars.
4. The invitation display duplicates each time you open a folder in a new tab or window.
   More work needs to be done to hook into the right events to create and remove the display properly.

To address 1,2,3 I intend to hide some information when too much, offering the
user the option to display when needed. For the description, I am considering
leaving that out all together in favor of the message body or keeping it really short,
displaying the excess in a modal, popup etc.

Differential Revision: https://phabricator.services.mozilla.com/D137583