Bug 1665511 - Fix focus handling in the spaces toolbar. r=aleca
Previously, when the spaces toolbar was revealed via the pinned tab button, or the status bar button, the focus was not moved into the spaces toolbar. This was because the code attempted to focus the spaces toolbar whilst it was still hidden, causing "focusButton.focus()" to fail.

We also ensure that focusButton follows a mouse click.

Differential Revision: https://phabricator.services.mozilla.com/D142849