Bug 1767509 - Make update-verify-config work better for Thunderbird releases. r=#thunderbird-reviewers,mkmelin
update_verify_config.py is mostly copied verbatim from gecko_taskgraph. The
changes are in INCLUDE_VERSION_REGEXES and MAR_CHANNEL_ID_OVERRIDE_REGEXES
which are modified to be more suitable for Thunderbird.

The "beta" INCLUDE regex will only match version numbers ending "bN". This
is to avoid trying to test updating a release version like 91.0 to a beta.
For Firefox, the "bN" is optional as they test updating release versions to
late betas.

Differential Revision: https://phabricator.services.mozilla.com/D145386

Depends on D145385