Bug 1695596 - "Update Failed" buttons open channel-specific URLs. r=mkmelin
When an update fails, the prompt has "Download Manually" and "More Information"
buttons. They opened pages on thunderbird.net that are suitable for stable
releases, but not for beta.

Move the prefs in question to branding, and use the update channel name to set
beta and release specific URLs.

For all non-shippable builds that use "nightly" branding, just set both URLs to
www.thunderbird.net as there is no "Daily" download page and there are no "Daily"
release notes. This will also be used for local developer builds.

The new "details" URL for beta needs a PR merged on the website before it will
work.

Differential Revision: https://phabricator.services.mozilla.com/D148323