Bug 1717632 - Split out the contact edit form into a distinct element. r=aleca
This improves the semantics and helps separate the name/nickname/email headings logic in aboutAddressBook.js from edit.js.

As a side effect, we also prevent dropping a contact photo whilst in viewing mode. Before this change, dropping outside of editing mode was 'responsive', but didn't actually save the photo, so we remove this functionality entirely.

Differential Revision: https://phabricator.services.mozilla.com/D148900