Bug 1777920 - Fix bad escaping of vCards by Google's CardDAV server. r=mkmelin Removes the extra escaping added to \r and : (and sometimes , ; and \) characters by the Google server.
The removal happens as a card is read from the database, so no changes are permanent (unless the card is edited, in which case Google will put the escaping back again) and existing cards don't need fixing.
Because it's done at this point, none of the UI needs to be aware of it.

Differential Revision: https://phabricator.services.mozilla.com/D151704