Bug 1775245 - Refactor tree-listbox selection code when selected item is removed. r=darktrojan a=wsmwk
When deleting a focused and selected item, we try and select the next item that was not deleted, or the last item overall. This prefers keeping the selection at the same position.

For nested items, we restrict the new selection to lie within the same remaining ancestor, or fallback to the ancestor itself. This prefers moving the selection to a neighbouring item under the same parent.

We also simplify the logic by using selectedRow (an Element) rather than selectedIndex: the latter would require careful update whenever items are collapsed, expanded, removed or added.

Differential Revision: https://phabricator.services.mozilla.com/D151390