Bug 1782374 - Add UrlListener to serialise IMAP folder expunge and compact operations. r=mkmelin
On IMAP folders, the user-facing concept of "compact" also implies an expunge.
This patch ensures that the expunge operation completes before the compact is started.
This means that expunged messages will be compacted immediately rather than being missed and left for the next compaction.
It also makes sure that any provided listener is properly informed when operation as a whole is finished (or has failed).

Differential Revision: https://phabricator.services.mozilla.com/D153827