Bug 1783441 - Remove X-LIC-ERROR handling in CalIcsParser.parseString. r=#thunderbird-reviewers,darktrojan a=wsmwk
I'm not 100% sure what the purpose of these checks are but I beleive X-LIC-ERROR is a libical specific thing.
Calling toString() on an icaljs freebusy property seems to mess up the formatting of the period string.

Geoff tracked the cause down to here:
https://searchfox.org/comm-central/rev/f57827565398fe6be1969d5107050822b5f2444f/calendar/base/modules/Ical.jsm#856-868

Not calling toString() while checking for X-LIC-ERROR avoids the problem though. Note that the check is not done in the
 async branch of parseString().

Differential Revision: https://phabricator.services.mozilla.com/D154571