Bug 1781072 - Add bouncer-location job for Thunderbird Daily builds. r=dandarnell
The download bouncer has new support for the "thunderbird-nightly-latest"
products, similar to the aliases that exist for beta and release. Users
will be able to use URLs like
https://download.mozilla.org/?product=thunderbird-nightly-latest-SSL&os=linux64&lang=en-US
which will always point to the latest Daily build.

The bouncer-locations job is responsible for updating the download URL post
merge-day version bump.

Differential Revision: https://phabricator.services.mozilla.com/D154840