Bug 1794156 - Remove mapit prefs and js. r=aleca
These prefs are not used anymore by Thunderbird (they are for Seamonkey, so they
are #ifdef wrapped in mailnews.js). The javascript is not loaded by Thunderbird,
but is still shipped.

Removing these prefs allows for cleanup of l10n.toml and should mean that the
remaining prefs will be exposed for localization in Pontoon again.

Differential Revision: https://phabricator.services.mozilla.com/D158869