Bug 1801150 - Set l10n locales-file to resolve based on release-type. r=dandarnell  a=wsmwk
The locales list needs to be different for release/beta builds, but that isn't
permitted by gecko_taskgraph. Some of the transforms for shippable-l10n need to
run before the gecko_taskgraph transforms, others after. This is handled by
creating two TransformSequence objects in a single file and arranging the kind
configuration accordingly.
The result is "locales-file" is now keyed by release-type.

Differential Revision: https://phabricator.services.mozilla.com/D162314