Bug 1803705 - Re-implement selectable column in message thread table header. r=darktrojan,freaktechnik
This patch only re-implements the selectable toggle in the table header.
I haven't implemented the selection toggles in the tree rows
to keep things a bit more simple to review. I'm already adding a lot
of things and I want to keep things smaller and quicker to updated
and land.

- Improved the UI of the columns and splitters.
- Implemented an initial threadPane object so we can clean up the code.
- Made the select toggle implementation agnostic.
- Various UX and UI improvements here and there.

Differential Revision: https://phabricator.services.mozilla.com/D164313