Bug 1823026 - Fix perma win comm/calendar/test/browser/eventDialog/browser_eventDialog.js. r=leftmostcat
This is a continuation to fixing the other issues in the attendees dialog.
For windows, one problem seems to be that on CI the screen on 1280 wide so buttons are pushed out of screen - which is not a problem on linux at least where dialogs are always capped at screen width.

But even with smaller width, windows has some focus issue. I don't know why that happens, but focusing the button before clicking it makes it work so I'll just leave it at that.

The start width is now smaller, the dialog can once again be resized to full width.

Differential Revision: https://phabricator.services.mozilla.com/D173226