Bug 1838758 - Use centralized cleanup() in config verifier. r=leftmostcat
There were cases where the temporary server was removed but not nulled out. Since it was not nulled out in those cases,
"onunload" handling removed a matching server which we did want to keep.

This occurred since no verification of password will take place if none was provided (and not OAuth2).
Normally after verification a proper cleanup step would take place, but since there was not verification,
this.server was left non-null.

Differential Revision: https://phabricator.services.mozilla.com/D181182