Bug 1848904 - Check a folder and its descendants are added to/removed from virtual folders when flags change. r=mkmelin
If a folder's flag changes, or a folder already with a flag is added or removed, it AND its
descendants should be added to the appropriate virtual folder. This patch tests that and fixes
some inconsistent behaviour.

In reality for most folder types this is a very rare occurrence, and for some it would never
happen, but it might as well work the same way for all types.

Differential Revision: https://phabricator.services.mozilla.com/D186327