Bug 1858896 - Port bug 1858006: Make installer auto-pin on Win 11 pre-UBR 2361 builds. r=rjl a=wsmwk
The previous commits to shared.nsh resulted in auto-pinning only occurring in Windows 11 builds that weren't build number 22621 or higher, but a lot of users on Windows 11 are on 22621 already and the build number to pay attention to, when MS introduced the user confirmation for pinning popup, is in Windows 11, 22621, UBR 2361.
This commit checks for the UBR number too.

Differential Revision: https://phabricator.services.mozilla.com/D190909