Bug 1857881 - Revert imap CLOSE functionality to ESR102 and earlier behavior r=mkmelin
Go back to doing imap CLOSE on folders with marked \deleted message so they are
expunged and thus can't appear as still present in other clients that maybe don't
respect the \deleted flag. This won't affect accounts configured as "just mark
message as deleted" as should be used when doing bulk folder moves between servers.
This effectively reverts most of the the changes to nsImapProtocol.cpp/.h
introduced in Bug 1828372 but it still allows inhibiting autoexpunge when the
user sets preference mail.imap.expunge_option to 3. It also uses a new name
for imap close function, ImapClose(), to avoid ambiguity with other close
methods in the huge .cpp  file.

Differential Revision: https://phabricator.services.mozilla.com/D190645