Bug 1868954 - [comm-esr115] Fix browser_applyView.js. r+a=rjl
This test picked up some errors when it was backported, not to mention the
default behaviour in 115 is different from that of Daily.
Most problematic is an earlier change to the test which was never uplifted
(because it's part of a much bigger patch not intended for 115):
https://hg.mozilla.org/comm-central/diff/3029c347bcfb57fad094ee55b43ff674a295c947/mail/test/browser/folder-display/browser_applyView.js

As well as fixing the test I've made some further changes to make it a better
test. I'll push these changes to comm-central at a later point.

Differential Revision: https://phabricator.services.mozilla.com/D195853