Bug 1881560 - Decode sub messages even if decodeSubMessages is set to false, but pruneat is set to a part in a sub message. r=mkmelin
Currently, it is not possible to get the source of a message (not
parsed), which is attached to an attached message (or any deeper nesting).

If I use `decodeSubMessages = true`, I get all the details of the deep nested
message, but not its raw untouched body. If I use `decodeSubMessages = false`,
I get the full source of the primary attached message. I need to parse that
message again with `decodeSubMessages = false`, to get the source of its
attached message. Or I have to manually plug the message back together
(including all its parsed sub parts).

It is much faster and less error-prone to dynamically enable
decodeSubMessages for a part, if the part pruneat is set to is *inside*
that part.

Differential Revision: https://phabricator.services.mozilla.com/D202463