Bug 1882352 - Set the DICPATH environment variable. r=dandarnell a=wsmwk
According to bz#1881830 the `spellchecker.dictionary_path` config
in distribution.ini has no effect and no spellcheckers other than
en-US show up. The same can be observed for Thunderbird flatpak and the
workaround listed in the above bug works for Thunderbird too.

So pass that environment variable in the flatpak

Differential Revision: https://phabricator.services.mozilla.com/D202855