Bug 1827594 - Open protocol_handler pages in the background, if opened from the console. r=mkmelin
Default to opening protocol-handler tabs in the background, if opened
via the console. Since this can be triggered externally (with
Thunderbird already running), the user might become distracted, if such
a tab suddenly opens and steals focus. If really necessary, the protocol
implementation itself can bring the tab into the foreground.

Differential Revision: https://phabricator.services.mozilla.com/D203096