Bug 1042344 - Show the close button in the cookies modal dialog on macOS. r=mkmelin
Historically the cookies popup didn't have a close button on macOS because (maybe) it was a standalone popup
dialog handled by the macOS titlebar and window decoration system independently.

Since this dialog is now (since 10 years) a modal dialog, it needs a close button as shown on Windows and Linux.

Differential Revision: https://phabricator.services.mozilla.com/D203810