Bug 1849512 - Create a LightweightThemeConsumer instance also in the message browser. r=mkmelin
We have an observer in `MailGlue.sys.mjs` for `chrome-document-global-created`,
to create a new instance of the `LightweightThemeConsumer` for each new
opened document.

This is not fired for the `messageBrowser`, which appears to be a
content document.

This patch makes sure we also create such an instance for the
messageBrowser. The patch also adds a test.

**The Bug itself has a reproducer add-on, showcasing the introduced
change by adding a border to all theme-able elements. Without this
patch, the message header area is not touched.**

Differential Revision: https://phabricator.services.mozilla.com/D204029