Bug 1804170 - Properly wait for action popups being closed. r=mkmelin
Another reason why this type of test could fail: We are not waiting till
the popup is closed but use a timeout.

This patch uses the helper `closeBrowserAction()` in all places,
which waits for the "popuphidden" event.

It also increases readability by using the  named event "popup closed".

In the `browser_ext_browserAction_popup_focus.js` test, the
wait-for-closed logic is moved from the extensions background to the
test itself, because it is much more stable this way.

Differential Revision: https://phabricator.services.mozilla.com/D204761