Bug 1804170 - Unify clicks on menu items and hiding menu popups. r=mkmelin
We know we have issues with clicks on menu items and hiding menu popups
in the WebExtension tests. We currently mitigate these issues not only
in header files, but also directly in tests, and we also still have
places which are not using the mitigation (which could be the reason for
this intermittent).

While I still attempt to find the underlying cause for these issues,
this patch tries to unify the mitigation.

1. Use the helper `clickItemInMenuPopup()` instead of manually calling
`menu.activateItem()` in the tests themselves. The helper waits for the
"popupHidden" event, which was not waited for in all places.

2. Use the helper `closeMenuPopup()` instead of manually calling
`menu.hidePopup() in the tests themselves. The helper waits for the
"popupHidden" event, which was not waited for in all places.

3. The function `closeContextMenu()` has been renamed to
`closeBrowserContextMenuPopup()` and no longer accepts a menu
parameter, but is now used only for the browserContext menu. If
a specific menu should be closed, `closeMenuPopup()` should be used.

4. The function `closeExtensionContextMenu()` has been renamed to
`clickItemInBrowserContextMenuPopup()`, which better describes what it
does.

Differential Revision: https://phabricator.services.mozilla.com/D204760