Bug 1885934 - Use BrowserTestUtils.waitForPopupEvent() instead of manually waiting for events. r=mkmelin
This is a continuation of the effort started in D204760: Switching to
BrowserTestUtils.waitForPopupEvent() instead of waiting for
`popuphidden` and `popupshown` events. It uses this general syntax:

```
await BrowserTestUtils.waitForPopupEvent(menu, "shown");
// Do something that should hide the menu.
await BrowserTestUtils.waitForPopupEvent(menu, "hidden");
```

or

```
await BrowserTestUtils.waitForPopupEvent(menu, "hidden");
// Do something that should show the menu.
await BrowserTestUtils.waitForPopupEvent(menu, "shown");
```

The main goal is to get rid of the artificial setTimeouts. This patch
also removes duplicated code and adds javadoc descriptions to the
touched functions.

Differential Revision: https://phabricator.services.mozilla.com/D205027