Bug 1758470 - Lock compose window during print preview. r=aleca,vineet
The open/close Promises of PrintUtils are not returned from the m-c
component and are not accessible. The only access handle I could find is
the browser itself. Attaching the two callback methods is a very simple
solution, but is it any good?

This patch locks the composer (for example menus, action buttons
and commands) while the print preview modal dialog is shown.

Differential Revision: https://phabricator.services.mozilla.com/D204553