Bug 1816623 - Attempt to fix browser_ext_menus_replace_menu.js. r=mkmelin
This patch checks all calls of browser.menus.create() and makes sure
the menu creation is correctly awaited, before the test continues.

I identified the following as the main reason for the observed
intermittent fails: Even though the event listeners for the contextmenu
events have been successfully added to the element inside the shadow
DOM, they are not active immediately. One can observe the original
context menu being opened, which should not happen. I did not find an
event to wait for, which properly checks this. I therefore use a small
delay.

Differential Revision: https://phabricator.services.mozilla.com/D205191