Bug 1890785 - Attempt to fix "can't access property querySelector, cardsList.getRowAtIndex(...) is null" in browser_contact_tree.js. r=mkmelin
I guess the issue is caused by openAddressBookWindow() and similar test
helper functions only waiting for the "load" event, but our front end
code has a lot of async UI code which is triggered on load and therefore
the test tries to access elements too early.

This uses a similar fix as introduced in D207481.

Differential Revision: https://phabricator.services.mozilla.com/D207672