Bug 1738611 - Fix Convert To > Event not saved on some CalDAV servers due to illegal CR in DESCRIPTION r=mkmelin
The content of emails may contain an \r\n.
This text is always converted from \r\n to \n when editing in the event/task. However, if the text is not changed, \r\n is passed on, which leads to a broken ical later on. We should therefore remove \r from the start.

Differential Revision: https://phabricator.services.mozilla.com/D207798