Bug 1847137 - Run POP3 accounts serially instead of in parallel. r=mkmelin
When a user has multiple POP3 accounts, instead of trying to run
them all at the same time, this causes them to be run sequentially
so that staggering the check for new message time is not needed
to avoid slow-downs/freezes.
This also changes the logging "prefix" to include the account's
server key (e.g., "server5") along with the run sequence number to
make it easier to distinguish each account's logged entries. E.g.,
prefix is now "pop3.server5.22": info for server 5, 22nd check
for new mail during session.
This still contains serveral dump calls for tracing that will be
removed before final approval. There are also some overly verbose
comments that explain the changes that probably need editing/removal
but are left in for now.

Differential Revision: https://phabricator.services.mozilla.com/D206738