Bug 1899314 - Rename MailTab.viewType to MailTab.groupType in Manifest V3. r=aleca
This is a cosmetic change. I implemented `viewType` as one of my first
tasks for the Thunderbird project and naming things is always difficult.
It turned out (and I was correctly called out on it), that the name is
not correct. The property which can be set to `ungrouped`,
`groupedByThread` and `groupedBySortType` should be named `groupType`.

I would like to correct that for Manifest V3.

Differential Revision: https://phabricator.services.mozilla.com/D211861