Bug 1852746 - Fix listAttachments() being empty for encrypted messages. r=mkmelin
This is a regression from switching to the MimeTreeEmitter. The old
implementation always tried to decrpypt the message for attachment
extraction. The new implementation did not, which is fixed by this patch.

This also adds an extended test to actually check listAttachments() on
an encrypted message (including nested messages).

Note: getRaw() and getFull() have an option to work on the original or
on the decrypted message. This patch does not add the same option for
listAttachments(). We currently do not handle encrypted parts as
"attachments" (this is why listAttachments() did not return the
raw encrypted part before this patch) and there has not been a request
to access the raw enrcrypted part. We may come back to this at a later time.

The changes in `utils.js` are fixing an incomplete type check. Only
`boolean`, `number` and `string` variables where type checked,
but not `date` for example. This patch also adds some more verbose
output to make it easier to track errors.

Differential Revision: https://phabricator.services.mozilla.com/D212990