Bug 1901930 - Fix inconsistency regarding messageDisplay.getDisplayedMessages() and mailTabs.getSelectedMessages() in Manifest V3. r=mkmelin
We introduced message lists in order to cope with large amount of
returned messages. We should be consistent in its usage. The number of
displayed messages can be as large as the number of selected messages,
so both functions should behave the same.

The only other place where we still return messages as plain arrays is
in the compose API, when returning the saved/send messages, which is
limited to only a few messages. It is acceptable to keep returning these
in an array.

This change is only affecting Manifest V3.

Differential Revision: https://phabricator.services.mozilla.com/D213384