Bug 1900790 - Add mailTabs.getSelectedFolders() and ensure correct values for "selected" folders in the menus API. r=aleca a=wsmwk
This adds the discussed method `mailTabs.getSelectedFolders()`, which
always returns the actual selected folders in the folder pane. The
menus API has been adjusted to always return the folders, which should
be acted upon by the context action (which might just be the clicked
folder, if it is not part of the current selection), following
Thunderbirds own UI.

Adding a test to verify both behaviors.

Differential Revision: https://phabricator.services.mozilla.com/D212688