Bug 1902221 - Allow to specify multiple accounts and folders for messages.query(). r=welpy-cw
There are reported cases, where it would be helpful to specify multiple
accounts or folders to set the search order, or search multiple
accounts without having to search each account individually.

Instead of deprecating the `accountId` and `folderId` properties in
favour of `accountIds` and `folderIds` properties, this patch adds the
option to either specify a single value, or an array.

The reasoning is that the search sets a criterion for a property of
the message, the `accountId` or the `folderId` in this case. We did the
same thing in `tabs.query()` and the search for the `type` property. See

https://searchfox.org/comm-central/rev/7ae782ce03a137bfd8674fc1cdcf10edfc1e6fb4/mail/components/extensions/schemas/tabs.json#440-451

Differential Revision: https://phabricator.services.mozilla.com/D213540