Bug 1906833 - Fix compose.getComposeDetails() if the "encryption" button has been removed. r=mkmelin
The purpose of checking the actual button state is to make sure to
report enabled encryption, even if no encryption technology is available
anymore due to an identity change or other actions performed by the
user. In this case, we still want to report enabled encryption, if the
user had it enabled. This is an invalid state and the message will not
be send (because encryption is not possible).

To fix the reported bug, we fallback to a state flag, if the button has
been removed. That state flag should be identical to the button state,
but Thundebird's security engineer explicitly wanted to check the button
state, so we keep doing that, if the button is available

Differential Revision: https://phabricator.services.mozilla.com/D216100