Bug 18732 - Add the linux system tray icon crate and its ksni dependency, r=aleca,babolivier,heather
This initial version of the system tray crate is geared specifically for use on Linux
via the Status Notifier Item specification. However in the interests of collapsing similar
codepaths we've left just enough space to allow expanding this into a more general purpose
entry point for all supported platforms.

Note that it was very difficult to support x86 (32-bit) builds due to needing build environment changes
with regards to libpthread, so at this point the feature is explicitly requiring a 64-bit build of Thunderbird.
This can be alleviated in time with future iterations to ksni and experimenting with
`zdbus` to drop the `libdbus-sys` dependency, opening up the door to many other D-BUS
services in Thunderbird.

Note that this is a collaborative effort between Heather Elssworth and myself,
with Heather having completed the original groundwork and research. I focused more
on the integration aspects and threading model. This feature wasn't possible without
Heather's work.

Differential Revision: https://phabricator.services.mozilla.com/D215249