Bug 1911752 - Update WebExtension schema files to improve documentation. r=aleca
The schema should be explicit about enum strings in the menus API to
improve validation results. Schema files should also be more explicit
about the expected content of email header strings.

This patch also removes a wrong information from the mailTabs API (there
is no `selectedMessage` property).

Furthermore, the different implementations for the space buttons for
Manifest V2 and Manifest V3 in the spaces API have been renamed. The
names were originally chosen as a reminder to find a different solution
(have a common type with individual changes in the type definition), but
that turned out impossible. Since these names look strange in the
documentation, we now call the type `ButtonProperties` in Manifest V2
and `SpaceButtonProperties` in Manifest V3. This change as no effect on
add-ons code.

Differential Revision: https://phabricator.services.mozilla.com/D218622