Bug 1912022 - Fix _execute_browser_action command in menus API. r=aleca  a=coreycb
Found to be not working while investigating bug 1911752. Will add a
dedicated test for this in bug 1912021.

This seems to be a left-over from the time when the mail tab was split
into sub windows (or it never worked). A mouse click event on the folder
pane or message pane is passing the `about:3pane` window and the
detection code for the unified toolbar (see [1]) fails. The action is
therefore not triggered.

[1]: https://searchfox.org/comm-central/rev/4010d287c4585532f986b886f4bccad478a1455a/mail/components/extensions/parent/ext-browserAction.js#61

Differential Revision: https://phabricator.services.mozilla.com/D218731