Bug 1915259 - Fix inconsistent value (null vs. undefind) in MailFolder.subFolders. r=aleca a=coreycb
This fixes a bug, but will change extension behavior (this is why a
test has to be changed).

A `MailFolder` returned as the `rootFolder` of an account will have its
`subFolders` property be `null`, if subfolders are not requested. All
functions from the folders API return `undefined` in this case.

The `rootFolder` property which is changed was only introduced recently
in Thunderbird 121 (Bug 1863269). I evaluated the chance to permanently
break add-ons to be very low, because the feature can only be used by
add-ons which have been updated recently and therefore have an active
maintainer. I will monitor extensions and help developers to mitigate
the change.

Extensions are affected, if they explicitly check for

`rootFolder.subFolders === null`

I did not find affected add-ons on ATN.

Differential Revision: https://phabricator.services.mozilla.com/D220337