Bug 1901962 - Test message sending scenarios. r=mkmelin
These test sending messages from the compose window with a variety of connection/authentication
problems. They are closely related to the tests named browser_getMessages with the same suffixes.

Sending a message presents a variety of prompt windows to the user, particularly when something
goes wrong. I've tried to avoid writing tests that lock in this behaviour in the hope that one day
somebody will overhaul it.

Differential Revision: https://phabricator.services.mozilla.com/D220857