Bug 1909879 - Fix flatpak crash caused by failure to register systray bus name, r=ikey a=coreycbryant
As a sandboxed app, it is better to implement StatusNotifierItem on the bus
name we already own rather than on a well-known bus name that will clash
(due to predictable PIDs inside sandboxes) or be denied.

The bus name being denied by default caused Thunderbird to crash
when run from flatpak.

More info here:
https://bugzilla.mozilla.org/show_bug.cgi?id=1909879#c7
https://issues.chromium.org/issues/40253493

Differential Revision: https://phabricator.services.mozilla.com/D219037