Bug 1882832 - Process encrypted MIME parts based on headers, minimize reading MIME parts. r=kaie a=coreycb
This is a performance improvement based on Kai's input to minimize the
computational work required to determine whether a message is encrypted.

This also fixes a bug where decrypted attachments where not
returned to the API (but the encrypted one), because a status flag was
not properly set.

This also adds tests.

Differential Revision: https://phabricator.services.mozilla.com/D202755