Bug 1905941 - Merge automation workaround for comm-central-to-beta and .gecko_rev.yml. r=dandarnell
This should be fixed by adding functionality to deal with .gecko_rev.yml to
Treescript. However in the interest of getting merge automation working add
this workaround.
There's no GECKO_HEAD_REV field in .gecko_rev.yml when comm-central-to-beta
runs, so instead create a line that can be easily replaced with the correct
data.

Differential Revision: https://phabricator.services.mozilla.com/D221194