Bug 1908528 - Allow source docs to be located outside of /docs path. r=dandarnell
This is handled by the rocsphinx Sphinx extension, a fork of mozsphinx. It
differs from mozsphinx in a key way: The document paths are not extracted from
moz.build files, they are set in `docs/config.yml` in the `doc_trees` mapping.
This was done to keep rocsphinx from importing code (mozbuild) that is in
the Gecko tree. When ReadTheDocs builds the documentation, it only has the
Comm source to work with.
The config file also has a `python_package_dirs` list structure which can be
used in the future to include Python documentation for comm_taskgraph among
other things.

Differential Revision: https://phabricator.services.mozilla.com/D221054