Bug 1929868 - Port bug 1859597: Do not compile msgMapi.idl during L10n repackaging. r=arschmitz,dandarnell
Bug 1859597 changes desktop_l10n.py in a way that runs `make export`
recursively through the object directory. The MAPI integration for Thunderbird
has an IDL file that the build system tries to compile as a result. It's not
needed for L10n repackaging, so exclude it when --disable-compile-environment
is set in mozconfig.

Differential Revision: https://phabricator.services.mozilla.com/D228347