Bug 1931424 - Add pref_false in app notification filtering criteria. r=tobyp
This assumes unset prefs are true. This makes it so pref_false isn't just the opposite of pref_true and
allows us to craft a targeting statement that would target an unset pref:

{
  exclude: [
    { pref_true: ["sample.unset.pref"] },
    { pref_false: ["sample.unset.pref"] }
  ]
}

This works by excluding any profile where the pref is explicitly set to true and any profile where it is
explicitly set to false. As such, we cover the entire tri-state of a boolean pref with pref_true and
pref_false.

As such, to target a profile that has a pref either unset or set to false, pref_false is the wrong tool,
instead pref_true with an exclude should be used. Similarly, to target a pref being true or unset
pref_false with an exclude should be used.

Differential Revision: https://phabricator.services.mozilla.com/D230520