Bug 1905616 - Open same-site links with a target attribute on http(s):// content pages in a new tab, if required. r=mkmelin
We already open same-site links with a target attribute on
moz-extension:// content pages in a new tab if needed. We do not have
special code for that, but this is correctly handled by the default
click handler.

The default click handler (which I was not able to find), however does
not handle this correctly for http(s):// content pages. The page is
opened in the same tab and then goes blank.

This patch is adding code to check if the requested BrowsingContext
exists and can be navigated to. If not, the default handler action is
prevented and we manually open a new tab.

Differential Revision: https://phabricator.services.mozilla.com/D229753