Bug 1941211 - No longer remove search address books when the background page is terminated. r=aleca
This event creates a read-only address book when being registered, and
removes it again, when being unregistered.

Manifest V3 terminates inactive backgrounds to save resources, which
causes the mentioned address book to be removed. This patch adds support
for the event to be registered as a persistent listener, which will
immediately re-register the listener as a "primed" listener, when the
background terminates. Instead of removing and re-adding the address
book while the listener is primed, this patch keeps and re-uses the
existing address book, to reduce flicker in the UI.

It was necessary to move the actual listener implementation out of the
`ExtSearchBook` class, into the closure of the event definition, to be
able to use the updated `fire` function.

Differential Revision: https://phabricator.services.mozilla.com/D234495