Bug 1898585 - Auto-generate tag key if not provided. r=aleca
Currently, `messages.tags.create()` requires specifying a tag name and a
tag key. While this provides maximum flexibility, there are also
scenarios where the add-on has no need to come up with a specific key or
actually wants to use the exact same key that would be used if the tag
is created through the UI.

This patch makes the key parameter optional and returns the key
associated with the created tag.

Differential Revision: https://phabricator.services.mozilla.com/D235002