Bug 1913621 - Allow extensions to control where links should be opened (in Thunderbird or in the user's default browser). r=aleca,mkmelin
We fixed many of Thunderbird's link issues in Bug 1905616, and we can
now offer to handle links by Thunderbird, even if they go to different
hosts. This is for example needed for add-ons utilizing the spaces bar
to load intranet pages, which might be spread across different domains.

This does not mean that Thunderbird can now be used as a browser. We
lack support of many web features and implementing those does not
align with Thunderbird's overall strategy.

This patch adds a `linkHandler` property to the create() functions of
the tabs API, the windows API and the spaces API. The new property can
be set to `balanced` (links to same host are opened in Thunderbird,
everything else in the user's default browser), `strict` (only links to
the same page are opened in Thunderbird) or `relaxed` (all links are
opened in Thunderbird).

Differential Revision: https://phabricator.services.mozilla.com/D232068