Bug 1925825 - Part 1 of 2 - Tweak nsMsgDBService::OpenMailDBFromFile() to use db filename rather than folder path. r=thunderbird-back-end-reviewers,darktrojan
The whole point of OpenMailDBFromFile() is that you can use it to access a db
in an arbitrary file, so it's a bit silly having to pass in a folder path.
It still only creates nsMailDatabase objects (used for local folders, not for
imap/news/etc), but that's coming in part 2.

Differential Revision: https://phabricator.services.mozilla.com/D237675