Bug 1947134 - Make sure the "tree-view" custom element is available before initializing the mail3pane. r=tobyp
The intermittent fail happens when `folderPane.selection` is
`undefined`.  However, `selection` is a getter for an internal Map,
which should be available as soon as a new instance of the module is
created.

I cannot find a scenario where this could happen, except when the
custom element is not yet defined. We do wait for many other custom
elements, but not for `tree-view`.

This patch therefore makes sure that the `tree-view` custom element is
defined, before executing the init code.

Differential Revision: https://phabricator.services.mozilla.com/D238945