Bug 1931277 - c-c part of using only interesting probes. r=tobyp
Only collect probes we have an interest in. The files I've is IMHO roughtly it. If we really want we can have specific files where we copy in only specific metrics - but the whole metric definition needs to be included (due to json schema validation).
The probes we do not care about are still enabled during tests, by applying interesting_serverknobs.json during automation runs, so that tests succeed. This work was enabled in bug 1942797.

Differential Revision: https://phabricator.services.mozilla.com/D228972