Bug 1955945 - Fix extra notifications in browser_notification test due to missing cleanup. r=darktrojan
Because test_no_notification_for_uninteresting_folders created a folder with new messages that it never
cleaned up, we ended up having multiple notifications show for the following tests. To resolve this, the task
now cleans up the folder it added. The alert service mock also enforces that it only gets one new notification
per cycle. Not only does this prevent weird states like we see in the failure, but it also helps us enforce
some correctness in the tests, since we usually don't check we're getting the correct notification, we just
check we're getting some notification.

The other changes are fixes to the cleanup to make this test run with --verify, making it run faster by
waiting for a promise instead of waitForCondition and lastly skipping the newmessage.xhtml task for headless
runs, since it doesn't work there.

Differential Revision: https://phabricator.services.mozilla.com/D242804