Bug 1957923 - Wait to use built in notifications until after first run. r=arschmitz a=wsmwk
This makes it so we will effectively never use built in notifications in the first run of a profile. I think that's generally fine. If we for some reason can't update from the network on the first run, we probably want to wait to the second run with built in notifications anyway.

Differential Revision: https://phabricator.services.mozilla.com/D244156