Bug 1958052 - Use a more traditional maildir file naming scheme, rather than "Message-Id.eml". r=heather
Part of the process of decoupling nsIMsgPlugableStore from the database.
Using Message-Id for the filename is problematic as we need the database to
access it. Also, it's not very... maildir-y.
This patch aims to use the more traditional scheme outlined in:
https://cr.yp.to/proto/maildir.html

But we still need to use a ".eml" file extension for OS search integration,
which also isn't very maildir-y.

Differential Revision: https://phabricator.services.mozilla.com/D244477