Bug 1953100 - Remove nsIMsgWindow.StopUrls() stopped and unused variables. r=heather
Stop is primarily used from the "Stop" button that can be added to the toolbar, and in relation to a few other operations: search, filtering, and also affecting offline sync on demand.

This patch removes stopurls and related code.
I did not come up with a good replacement for stop running filters on demand, so I suggest removing that functionality.
The related filtering code is not great, so there's no good place to hang the stopper on. We *could* think of something, but
it does seem like a rather obscure use case especially as filters are usually running quickly.

Differential Revision: https://phabricator.services.mozilla.com/D241003