Bug 1956601 - Record references when syncing EWS messages. r=heather
EWS exposes a `References` property on messages, which is a representation of
the RFC822 `References` header, and is a list of message IDs (RFC822 style)
formatted in the way the message DB code already expects it - so we can set it
on the database entry as is.

It also exposes a similar representation of the `In-Reply-To` header, and in
theory we'd be concatenating both. However, it looks like the EWS server takes
care of creating a `References` header and populating it correctly regardless of
whether the original message included one, so we should be able to rely
exclusively on it.

Differential Revision: https://phabricator.services.mozilla.com/D245058