Bug 1966171 - Clean up GlobalPopupNotifications.sys.mjs to remove "Argument 1 is not an object" error. r=arschmitz
We forked toolkit's `PopupNotifications.sys.mjs` [1] as
`GlobalPopupNotifications.sys.mjs` [2], but missed an upstream change in
one of its dependencies, which causes the following error

```
ChromeUtils.getClassName: Argument 1 is not an object
```

when installing an add-on and acknowledging the required notification
popups. A test case which throws the error without this patch is the
following add-on:

https://addons.thunderbird.net/addon/openai-proofreader

STR:

* download the add-on's XPI file
* open Thunderbird's add-on manager
* click on the gear icon and select "install from file"
* select the downloaded XPI file
* click "Add" on the shown notification
* observe the error in the console
* click "Ok" on the second notification
* observe the error in the console

There are two reasons for the fail:

1. We register event listeners to directly call `PopupNotifications`,
which is a function (prototyped to behave like a class) defined in the
same file, instead of `ownerGlobal.PopupNotifications`, which is the
lazy getter defined in `messenger.js` [3] to the global *instance* of
that class. The constructor of course does not know what to do with the
passed in event and dies.

2. There is no reason at all to register these listeners, as those
events are already handled by `popupnotification.js` [4]. They could be
registered to *prevent* the default event handler, but we are also not
doing that.

This patch is also following upstream in how to handle the no-shadow
lint issue, to make it easier to find relevant code.

[1]:https://searchfox.org/mozilla-central/source/toolkit/modules/PopupNotifications.sys.mjs
[2]:https://searchfox.org/comm-central/source/mail/modules/GlobalPopupNotifications.sys.mjs
[3]:https://searchfox.org/comm-central/rev/73897d5732eac86ed38bd521ca6f4610a4411186/mail/base/content/messenger.js#48-69
[4]:https://searchfox.org/mozilla-central/rev/be684f13a4ce71e7ca5e0b354a0f56a665842922/toolkit/content/widgets/popupnotification.js#163-172

Differential Revision: https://phabricator.services.mozilla.com/D249113