Bug 1966848 - Change all timers in pop-up notifications code to use a global proccess timer, and a few other upstream updates. r=mkmelin
We forked toolkit's `PopupNotifications.sys.mjs` [1] as
`GlobalPopupNotifications.sys.mjs` [2], but missed upstream changes.
This patch ports changes from Bug 1872993 and a few others.

Some changes in Bug 1872993 have already been changed again. This patch
is therefore based on the most recent version of
`PopupNotifications.sys.mjs`. To review the changes, one could compare
the current upstream version with our patched version and see that there
are no differences regarding `timeShown` and the use of `Cu.now()`.

We do not implement `_extendSecurityDelay()`, since we do not support
fullscreen mode. I will have to check what PointerLock is, and if that
is something we need, port the related code as well.

[1]:https://searchfox.org/mozilla-central/source/toolkit/modules/PopupNotifications.sys.mjs
[2]:https://searchfox.org/comm-central/source/mail/modules/GlobalPopupNotifications.sys.mjs

Differential Revision: https://phabricator.services.mozilla.com/D249741