Bug 1921394 - Include special folders in the sort order calculation. r=mkmelin I deliberately left this bit until later as it wasn't relevant at the time. I guess now is later.
I've overhauled the comparator logic while here to make it a bit more readable. I fixed a bug
where we didn't update the ordinal member of a Folder object when moving the folder. And I added
a method for resetting the order of a folder's children to the natural order, in case the user
wants to do that after reordering in the UI.

Differential Revision: https://phabricator.services.mozilla.com/D250381