Bug 1969229 - Query the correct interface when syncing an outgoing server removal. r=mkmelin
While testing this, I've also realised the EWS outgoing server didn't behave
correctly when queried about its UID if it didn't have one already. It seems
like the expectation is for the `nsIMsgOutgoingServer` implementation itself to
generate one if none exist, whereas `EwsOutgoingServer` was expecting it to be
set from an outside source.

Differential Revision: https://phabricator.services.mozilla.com/D251687