Bug 1972633 - Prepare schema files for automated compat data generation. r=tobyp,aleca a=coreycb
This is the final step of the reorganization started in Bug 1970044 and
Bug 1971694.

The webext-schemas-repository is now able to automatically generate our
compatibility data [1]. The manually (not very well) maintained data in
the webext-docs repository [2] is no longer needed - the docs can soon
also be automatically generated.

The schema files need a slight adjustment for this to work: If an API
parameter has multiple choices and a new choice is added, the new choice
MUST be added last. The compat data generator cannot check the type/ref
or name of parameters (which can be renamed!), but can identify new
parameters only through their position. If a parameter has two choices
which were added in TB78 and a third choice was added in TB92, but added
first, the `version_added` attribution will be wrong.

This patch also adds documentation and a few best practice suggestions.

Additionaly, the patch fixes a few schema errors and adds missing
entries, which were found while testing the compat data generator.

[1]:https://github.com/thunderbird/webext-schemas-generator/commit/3f8b969d2c580b2c000a9dfcf32a926070380f42
[2]:https://github.com/thunderbird/webext-docs/blob/Beta-MV2/overlay/accounts.json#L10

Differential Revision: https://phabricator.services.mozilla.com/D254058