Bug 1971540 - Part 2: Handle EWS authentication failures. r=edicharry
Although most of the error handling for EWS had been architectured around
happening inside the C++ callbacks, I realised this simply cannot work for
authentication errors as we'll want to retry the request with a different
authentication header in some cases (e.g. when a new password is entered). As a
result, this patch reworks the handling of authentication errors to be done in
Rust, closer to the code that generates and sends requests.

As a side-effect, I ended up rewriting a bunch of the existing alerting code
from `nsImapIncomingServer` in Rust for convenience, while also exposing the
new code over FFI so the C++ IMAP implementation can eventually start using it.

I've intentionally descoped removing the `IEwsClient::Error` enum (and
associated processing via the `process_error_with_cb*` functions) from this
patch, since this requires updating a lot of callback classes which would
unnecessarily bloat this already large patch. I'll take care of this in a part
3.

Differential Revision: https://phabricator.services.mozilla.com/D255652