Bug 1982798 - Prevent extraneous empty lines from being added to Cargo.toml when regenerating it. r=tobyp
It's a simple fix but it took me some time to fully understand why this is
happening. `tomlkit.TOMLFile.read()` preserves whitespaces (including empty
lines) when parsing a TOML file (so it can reproduce them if it needs to turn
the `TOMLFile` into a string). However, our template adds two line breaks at the
end of the `[features]` section (to separate it from the following `[workspace]`
section). Which are then reproduced endlessly.

I'm not entirely sure why it started happening at the time it did, but I suspect
this might have something to do with
https://github.com/python-poetry/tomlkit/pull/421 or
https://github.com/python-poetry/tomlkit/pull/422 (although I'm not 100% certain
since it doesn't entirely match what I saw when I was fiddling with the
tomlkit source code to investigate this). Both these changes shipped in
tomlkit 0.13.3 which released in early June, which is a couple of weeks before
this issue started happening.

Anyway, the right thing to do in our case seems to strip the extra line breaks
since we already take care of adding them in the template.

Differential Revision: https://phabricator.services.mozilla.com/D260963