Bug 1985735 - Support repairing EWS folders. r=mkmelin
Folder repair for EWS is mostly the same as with other protocols (i.e. we wipe
the message database and (when relevant) repair the message store, then trigger
a resync of the folder's messages), but on top of this we also need to reset the
sync state token so the resync starts from the beginning.

Differential Revision: https://phabricator.services.mozilla.com/D263890