Bug 707165 - Improve 'account' column for messages with account key. r=mkmelin
Messages originally received by a Thunderbird POP3 account contain an
'X-Account-Key:' header. If an account for this key can't be found, or if it
currently maps to a typeless incoming server such as 'Local Folders' or 'Unified
Folders', display the verbose account key instead of the pretty name for the
incoming server.

Differential Revision: https://phabricator.services.mozilla.com/D270084