Bug 1901089: Disable date pickers when their checkboxes are unchecked. r=darktrojan
The updateDateCheckboxes function contains logic
supposed to disable the date/time pickers when the
related checkboxes are unchecked.

However, that logic wasn't working because
it tried to set a property directly on the ID of
the corresponding date/time picker element instead
of on the actual element.

Make this actually work by retrieving the
corresponding element and setting its
"disabled" property.

Also adjust logic to set the value a few lines
above to also use the element instead of its ID.

Differential Revision: https://phabricator.services.mozilla.com/D269683