Bug 1875428 - Explicitely set add-on requireBuiltInCerts prefs to false. r=mkmelin,freaktechnik,kaie
This patch disables pinning of ATN certificates for add-on installs and
add-on updates. It has become a major cause for issues related to add-on
updates and installs when users use third party security software or SSL
proxies.

As outlined in the associated bug, the security gain seems to be small
compared to the negative impact for affected users.

Differential Revision: https://phabricator.services.mozilla.com/D201827