Bug 2001632 - Treat parts with a content-id as attachments only if they also have a name. r=mkmelin
Usually the presence of a content-id header indicates that a part is an inline element
referenced from elsewhere in the message as a related part. The current implementation
however treats those parts as attachments, which should only happen if the part has a
name.

This patch modified two tests, which locked down wrong behavior. They both work
on the same test message:

https://searchfox.org/comm-central/source/mail/components/extensions/test/xpcshell/messages/invite.eml

This message has a multipart/related part with the following sub-part:

```
--=_related_36138BFC6A7B03EFC2258776003DB5C5_=↩
Content-Type: text/html; charset=windows-1257↩
Content-Disposition: inline↩
Content-ID: <FFFF__=4DBB0DE5DFAE33558f9e8a93df938690918c4DB@>↩
↩
<html>↩
 <body>↩
  <p>You have been invited to a meeting (RELATED)</p>↩
 </body>↩
</html>↩
↩
--=_related_36138BFC6A7B03EFC2258776003DB5C5_=--
```

This part is specified as `inline` and has no name. The former implementation
treated it as an attachment, excluded it's body from `getFull()`, which was wrong.

This patch does not create a regression for bug 1737532 (whose test is changed).
All parts are still returned, either as attachments or as inline text parts.

Differential Revision: https://phabricator.services.mozilla.com/D273796