Bug 2002937 - Link returned data to the sending window. r=mkmelin
This is strange. The converted data is stored per window and there
should not be a way to leak data from one `MsgOperationObserver` to
another.

We currently do not check the `tab`, which is send with the event.
The only mechanism I could see is that the first window sends all
its events twice, causing the listener waiting for the second
window to receive the data for the first window again.

This patch is adding the tab id to the message, causing the second
part of the test waiting for "his" events. If we get the data for
the first event twice, we should see that in the final report of
unexpected messages.

Differential Revision: https://phabricator.services.mozilla.com/D274389