Bug 1945489 - Replace nsMsgStatusFeedback in SearchDialog. r=mkmelin
Simply set the text and progress indicator in the status bar of the
'Search Messages' dialog directly, instead of using the `nsMsgStatusFeedback`
implementation via `nsIMsgWindow`. Removing the latter as well isn't possible
without changes to `nsMsgSearchSession::InterruptSearch()`, however.

Also disable the 'Clear' button while a search is running, since clicking it
at this point would put the search dialog in a non-functional state.

Differential Revision: https://phabricator.services.mozilla.com/D274275