Bug 1998426 - Follow-up: remove leftover unused method in the Graph client. r=jtracey
Looks like I forgot to remove a method from `graph_xpcom` after some
experimentation while working through some of the intricacies of D274595. I
thought `xpcom_method!` would tell me off for using it with a method that isn't
related to the interface that is being implemented, but on second thought it
probably can't.

Differential Revision: https://phabricator.services.mozilla.com/D275750