Bug 1977855 - Add sorting support for browser.messages.list(). r=freaktechnik
Add-ons can currently sort returned messages only after receiving all
of them. The time-consuming part of this operation is creating the
WebExtension MessageHeader objects and passing them across the process
boundary. It is therefore faster to sort messages before passing them
to the extension.

The extension may still need to wait for all messages, but it can
already process or display the ones it has received while waiting for
the remaining message pages.

Once the global message database is in place, entries could be sorted
directly during the lookup request, which would be even faster.

Differential Revision: https://phabricator.services.mozilla.com/D275057