Bug 2005900 - a11y: Set "aria-pressed" for "Keep duration" button. r=arschmitz
Set the "aria-pressed" state [1] for the
"Keep the duration when changing the end date" button
in the dialog to create/edit an event or task.

This causes it to be exposed as a toggle button on the accessibility
layer (see [2]), and the current state is made available to
assistive technology like screen readers.
This e.g. makes the Orca screen reader on Linux announce the
button as "toggle button" that's either "pressed" or "not pressed",
making clear whether or not the option to keep the duration is
currently active or not.
(Previously, it was only announced as a "button" without any state,
so it was unclear to screen reader users whether the option
was currently active or not.)

[1] https://w3c.github.io/aria/#aria-pressed
[2] https://w3c.github.io/aria/#button

Differential Revision: https://phabricator.services.mozilla.com/D276352